# resources/documentation/swagger.yaml
openapi: 3.0.3
info:
  title: BarenAD - products
  description: |-
    Это документация API для демонстрационно-боевого* проекта доски товаров.

    *- проект сделан для саморазвития, но под требования заказчика и работает на него по сей день.
  version: 1.0.0
servers:
  - url: https://products.kemerovo.barenad.ru
...
paths:
...
  /api/auth/logout/all:
    post:
      tags:
        - Auth
      summary: Разавторизоваться на всех устройствах
      description: Удаляет все авторизационные токены
      operationId: logoutAll
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string
                example: OK
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedExeption'
      security:
        - bearerAuth: []
  /api/profile/permissions:
    get:
      tags:
        - Profile
      summary: Права пользователя
      description: Получает текущие права пользователя
      operationId: profilePermissions
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: users.roles.create
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedExeption'
        403:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotEnoughPermissionsExeption'
      security:
        - bearerAuth: []
...
components:
  schemas:
    UnauthorizedExeption:
      type: object
      properties:
        id:
          type: integer
          example: 0
        message:
          type: string
          example: Пользователь не авторизован.
    NotEnoughPermissionsExeption:
      type: object
      properties:
        id:
          type: integer
          example: 6
        message:
          type: string
          example: Недостаточно прав администрирования.
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT